+++ workflows_definition/RELEASES.md
@@ -179,3 +179,11 @@ Based on:
 - Speakeasy CLI 1.431.0 (2.451.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.10.2] workflows_definition
+
+## 2024-11-09 00:28:02
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/workflows-definition.yaml
+- Speakeasy CLI 1.436.0 (2.455.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.10.3] workflows_definition